Refactor secret provider code to separate backend and provider implementation
